Widen bitmap fonts for proportional glyphs, un-bold script-l, and tune chooser/scale-bar readability
Font tables now use per-size proportional glyph widths (FontSmall/Font widened to
uint16_t/uint32_t rows) instead of a fixed cell, and script-l drops its synthetic
bold stroke (generator no longer fakes a bold weight -- see generate_font.py).
The electron glyph rollout extends to the PC and web renderers so all three
targets share the same "e-" superscript-minus glyph instead of a plain string.

Chooser menu text now fits the 240px panel (was overflowing at scale 2), blinks
between two colors instead of on/off, and sits in the lower part of the screen.
Scale bar labels round 1000 pm up to "1 nm" instead of a 4-digit pm count.
Orbital point brightness uses a gamma-shaped rank curve (levelFromRankFraction())
instead of a flat linear spread, biasing the cloud brighter without popcorn at
the cutoff. web_atom.py picks up a Black formatting pass.

+++ src/render/font.h
@@ -37,13 +37,13 @@ struct FontBase
 };
 
 /** Row storage for kFontSmall -- its widest glyph is under 8px. */
-using FontSmall = FontBase<uint8_t>;
+using FontSmall = FontBase<uint16_t>; // quick test: widened from uint8_t for DejaVu's wider glyphs
 
 /**
  * Row storage for kFontLarge -- its widest glyph is under 16px. Kept as the plain "Font"
  * name since it's the size generic call sites (chooser.cpp, ticker.h) are written against.
  */
-using Font = FontBase<uint16_t>;
+using Font = FontBase<uint32_t>; // quick test: widened from uint16_t for DejaVu's wider glyphs
 
 /** Row storage for kFontHuge -- its widest glyphs (e.g. '#', '@') run past 32px. */
 using FontHuge = FontBase<uint64_t>;
